text: update
After more than a year I decided to re-write the reference with all of
the updates I've come up with during the development of the compiler.

The Reference is far from completeness, but it covers almost everything
what's in my mind. It's still too official. Shall make it more friendly